django_project/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-paypal==2.0
gunicorn==20.1.0
idna==3.3
migrate==0.3.8
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
stripe==3.0.0
urllib3==1.26.9
